Fix : - Add userId with jwt token and UserDetails - Fixes error of business logic of cart, add, get, update and remove functionality. - Fixes Order service business logic. - Fixes payment service business logic

+++ src/main/java/com/ztrios/Ecommerce/dto/res/CartItemResponse.java
@@ -1,11 +1,13 @@
 package com.ztrios.Ecommerce.dto.res;
 
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.UUID;
 
 public record CartItemResponse(
         UUID productId,
         String name,
         Integer quantity,
-        Double price
+        BigDecimal price
 ) {}
